+++ lib/__init__.py
@@ -2,7 +2,7 @@
 Library package for ACM switchover automation.
 """
 
-__version__ = "1.5.14"
+__version__ = "1.5.15"
 __version_date__ = "2026-03-30"
 
 from .exceptions import (